+++ weatherAppTest/Model/WeatherModel.swift
@@ -1,0 +1,25 @@
+import Foundation
+
+struct Weather: Codable {
+    
+    let weather: [WeatherModel]
+    let main: Main
+    let name: String
+    
+    struct WeatherModel: Codable {
+        let description: String
+    }
+    
+    struct Main: Codable {
+        let temp: Float
+        let tempMin: Float
+        let tempMax: Float
+        
+        enum CodingKeys: String, CodingKey {
+            case temp
+            case tempMin = "temp_min"
+            case tempMax = "temp_max"
+        }
+    }
+}
+
